Seperate goal utilities.
